refactor: split gas-analytics helpers.ts into helpers/ directory to break circular dependency
