+++ packages/hardhat/src/internal/builtin-plugins/gas-analytics/hook-handlers/test.ts
@@ -1,6 +1,6 @@
 import type { HookContext, TestHooks } from "../../../../types/hooks.js";
 
-import { getGasAnalyticsManager } from "../helpers.js";
+import { getGasAnalyticsManager } from "../helpers/accessors.js";
 
 export default async (): Promise<Partial<TestHooks>> => ({
   onTestRunStart: async (context, id, next) => {
